RegionSearchItem -> SearchItem, into its own folder
