+++ packages/ui-components/src/components/SearchItem/index.ts
@@ -1,0 +1,1 @@
+export * from "./SearchItem";
